+++ client/tests/addons/reviewer-report-test.js
@@ -30,7 +30,7 @@ module('Integration: Reviewer Report', {
     app = startApp();
     server = setupMockServer();
     fakeUser = window.currentUserData.user;
-    TestHelper.handleFindAll('discussion-topic', 1);
+    TestHelper.mockFindAll('discussion-topic', 1);
 
     records = paperWithTask('ReviewerReportTask', {
       id: taskId,