+++ client/tests/factories/question-attachment.js
@@ -1,0 +1,11 @@
+import FactoryGuy from 'ember-data-factory-guy';
+
+FactoryGuy.define('question-attachment', {
+  default: {
+    filename: 'test.jpg',
+    src: 's3/test.jpg',
+    status: 'done',
+    title: 'Test file',
+    caption: 'A test file'
+  }
+});